Split DOM related implementation to its own plugin
