+++ sbt-scalajs-esbuild-dom/src/sbt-test/basic-dom-project/src/main/scala/example/Main.scala
