+++ sbt-scalajs-esbuild-dom/src/sbt-test/basic-dom-project/src/test/scala/example/MainSpec.scala
