+++ sbt-scalajs-esbuild-dom/src/sbt-test/basic-dom-project/test
